requirements.txt
# Core Flask and Extensions
Flask==2.3.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# AI / ML
joblib==1.3.2
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.8.0

# API / JSON
requests==2.31.0

# Security
Werkzeug==2.3.7

# Utilities
geopy==2.4.0
python-dateutil==2.8.2
